requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.13.2
nltk==3.9.1
streamlit==1.45.1
openai==1.86.0
requests==2.32.3
python-dotenv==1.1.0
joblib==1.4.2
wordcloud==1.9.4
